+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
  * Adds a check for acceptable client certificate CAs
  * Supporting certificate expiration after 2038-01-19 on 32 bit systems
+ * Adds an option (```--ignore-connection-problem```) to set a custom state in case of connection failures